Tuareg and OCamlformat in opam/share/emacs/site-list
Tuareg and opam don't need to be installed anymore through emacs
The load-path is augmented with the current opam switch.
Currently, if you switch to a project with a different opam switch,
the switch will be automatically evaluated for compilation but not for
loading Ocamlformat since it has already been loaded. If you use
project with different versions of OCamlformat, just start a new emacs sessions.