Add code folding
This also works with tuareg even though it doesn't handle tree-sitter

Use S-tab to toggle folding (or simply tab on a folded node)

I added code folding for OCaml 3 years ago and I need to make a new PR to
improve it. Functors are folded in a strange way and module type definitions are
not folded. But hey, it works!

You may have to pokemacs-fix-treesit-auto ocaml the first time
(same with elisp if any issue arises)