OCaml LSP update (must read!!!)
I'm working on improving the OCaml LSP client in lsp-mode to provide an
experience on par with the one provided by the VSCode client and eglot client.

This is still a work in progress but I added the ocamllsp/typeEnclosing request
allowing to destruct aliases and copy types directly. If you pull this commit
you'll need to remove the elpaca/(build|repos)/lsp-mode to make sure that the
proper one is installed until my PR are merged. Any feedback is appreciated but
do it with "les formes" please <3